requirements.txt
python-dotenv
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
numpy
google-generativeai
pypdf
